Rename auto versioning module for clarity
The resolveVersion function lives in util, but there was a module named versioning. It was purely auto-version functions.
